Add check for too long cipher suite names
It might happen in future, that the cipher suite name that shall be
copied to u.crypt.alg is to long to fit in there. In this cases the
algorithm will proceed with what fits in but print a warning to the
system logs.

This behaviour will lead to failing WLAN connections but allow the user
to continue working and to diagnose the problem.